feat: call api for community detail

+++ backend/controllers/community.controller.js
@@ -1,4 +1,5 @@
 import CommunityService from '../services/community.service.js';
+import ApiError from '../utils/api_error.util.js';
 
 
 // Lấy tất cả cộng đồng
@@ -36,6 +37,10 @@ export const getCommunityById = async (req, res, next) => {
 // Tạo cộng đồng mới
 export const createCommunity = async (req, res, next) => {
     try {
+        // Nếu không phải admin, không được tạo mới cộng đồng
+        if (req.user.role.roleName !== 'admin') {
+            throw new ApiError("Bạn không có quyền tạo cộng đồng")
+        }
         const data = req.body;
         const community = await CommunityService.createCommunity(data);
         res.status(201).json(community);
@@ -47,6 +52,10 @@ export const createCommunity = async (req, res, next) => {
 // Cập nhật cộng đồng
 export const updateCommunity = async (req, res, next) => {
     try {
+        // Nếu không phải admin, không được cập nhật cộng đồng
+        if (req.user.role.roleName !== 'admin') {
+            throw new ApiError("Bạn không có quyền cập nhật cộng đồng")
+        }
         const communityId = req.params.communityId;
         const data = req.body;
         const community = await CommunityService.updateCommunity(communityId, data);
@@ -59,6 +68,10 @@ export const updateCommunity = async (req, res, next) => {
 // Xóa cộng đồng
 export const deleteCommunity = async (req, res, next) => {
     try {
+        // Nếu không phải admin, không được xóa cộng đồng
+        if (req.user.role.roleName !== 'admin') {
+            throw new ApiError("Bạn không có quyền xóa cộng đồng")
+        }
         const communityId = req.params.communityId;
         const result = await CommunityService.deleteCommunity(communityId);
         res.status(200).json(result);
